requirements.txt
pillow == 5.2.0
scikit-image == 0.14.2
tensorflow-gpu==2.0.0rc0
pandas == 0.24.0
bitstring == 3.1.5
matplotlib == 3.0.2
setproctitle == 1.1.10
